If there's no id to delete we don't know what to do yet, so handle generic

+++ app/controllers/inbox_controller.ts
@@ -79,6 +79,10 @@ const handleDelete = async ({ request, response }: Pick<HttpContext, 'request' |
   const body = request.body()
   const idToDelete = body.object.id
 
+  if (!idToDelete) {
+    return handleGeneric({ request, response })
+  }
+
   let note = await Note.find(idToDelete)
   if (note) {
     await note.delete()
